Renamed namespace to Archery
